backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "aiobotocore (>=2.21.1,<3.0.0)",
    "fastapi-cache2[redis]>=0.2.2",
    "python-ffmpeg>=2.0.12",
    "opencv-python>=4.11.0.86",
    "whisperx==3.3.2",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
    "hf-xet>=1.0.2",
    "torch>=2.6.0",
    "onnxruntime-gpu>=1.21.0",
    "argostranslate>=1.9.6",
    "webvtt-py>=0.5.1",
    "coqui-tts>=0.24.3",
]


[tool.uv.sources]
python-ffmpeg = { git = "https://github.com/Supun3746/python-ffmpeg.git" }
torchvision = {index = "pytorch-cu126"}
torchaudio = {index = "pytorch-cu126"}


[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[dependency-groups]
lint = [
    "ruff>=0.11.5",
]
